refactor : 팀선택 페이지에서 StepConfig 타입이 StepSlider에 있던 거 types/step.ts로 옮김
