+++ frontend/src/pages/teamSelectPage/types/step.ts
@@ -1,0 +1,6 @@
+import type { ReactNode } from 'react';
+
+export interface StepConfig {
+  label: string;
+  content: ReactNode;
+}
